selinux-testsuite: inet_socket: test xfrm state selectors
Augment the inet_socket ipsec-load configuration and test script
to test xfrm state selectors.  This isn't really testing SELinux
functionality per se but is rather illustrating the lack of xfrm
selector matching in the xdst pcpu cache.

I don't know if we truly want this in the selinux-testsuite; I used
it to confirm that we are not getting proper xfrm state selector
matching with the current xdst pcpu cache code and to test a possible fix.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>