selinux-testsuite: add testing for unprivileged sandboxing capability
This patch provides tests for the unprivileged sandboxing capability
introduced by the preceding kernel and userspace patches. In order for
these tests to work, the patched kernel and selinux userspace need to be
compiled and installed on the system.  This is an RFC only, see the
kernel patch description for a full description of the new capability.

To run these tests, in addition to installing the patched kernel
and userspace, you will need to modify your base policy as follows:

$ sudo semodule -cE base
$ sudo vi base.cil
<add "sandbox" to (class security (....))>
$ sudo semodule -i base.cil

$ cat ./definesandboxclass.cil
(class sandbox (load_policy unload_policy))
(classorder (unordered sandbox))
$ sudo semodule -i definesandboxclass.cil

$ sudo vi /usr/share/selinux/devel/include/support/all_perms.spt
<add "sandbox" to all_security_perms definition, and add the following
within the all_kernel_class_perms definition:
class sandbox { load_policy unload_policy };

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>